Pull the shared logger out to live in command, break up into files
This commit does not change any functionality from the last commit,
but moves things around to make them easier to find and reason about.
It additionally moves the logger that is used everywhere to the
command struct so it doesn't have to be passed into all the functions.
