+++ subcommand/server-acl-init/command.go
@@ -4,19 +4,17 @@ import (
 	"context"
 	"errors"
 	"flag"
+	"fmt"
 	"os"
-	"strings"
 	"sync"
 	"time"
 
-	"fmt"
 	"github.com/hashicorp/consul-k8s/subcommand"
 	k8sflags "github.com/hashicorp/consul-k8s/subcommand/flags"
 	"github.com/hashicorp/consul/api"
 	"github.com/hashicorp/consul/command/flags"
 	"github.com/hashicorp/go-hclog"
 	"github.com/mitchellh/cli"
-	apiv1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -61,6 +59,9 @@ type Command struct {
 	// cmdTimeout is cancelled when the command timeout is reached.
 	cmdTimeout    context.Context
 	retryDuration time.Duration
+
+	// Log
+	Log hclog.Logger
 
 	once sync.Once
 	help string
@@ -185,13 +186,13 @@ func (c *Command) Run(args []string) int {
 	// The context will only ever be intentionally ended by the timeout.
 	defer cancel()
 
-	// Configure our logger.
+	// Configure our logger
 	level := hclog.LevelFromString(c.flagLogLevel)
 	if level == hclog.NoLevel {
 		c.UI.Error(fmt.Sprintf("Unknown log level: %s", c.flagLogLevel))
 		return 1
 	}
-	logger := hclog.New(&hclog.LoggerOptions{
+	c.Log = hclog.New(&hclog.LoggerOptions{
 		Level:  level,
 		Output: os.Stderr,
 	})
@@ -199,7 +200,7 @@ func (c *Command) Run(args []string) int {
 	// The ClientSet might already be set if we're in a test.
 	if c.clientset == nil {
 		if err := c.configureKubeClient(); err != nil {
-			logger.Error(err.Error())
+			c.Log.Error(err.Error())
 			return 1
 		}
 	}
@@ -224,23 +225,23 @@ func (c *Command) Run(args []string) int {
 		}
 		return fmt.Errorf("rollout is in progress (CurrentRevision=%s UpdateRevision=%s)",
 			statefulset.Status.CurrentRevision, statefulset.Status.UpdateRevision)
-	}, logger)
+	})
 	if err != nil {
-		logger.Error(err.Error())
+		c.Log.Error(err.Error())
 		return 1
 	}
 
 	// Check if we've already been bootstrapped.
 	bootTokenSecretName := c.withPrefix("bootstrap-acl-token")
-	bootstrapToken, err := c.getBootstrapToken(logger, bootTokenSecretName)
+	bootstrapToken, err := c.getBootstrapToken(bootTokenSecretName)
 	if err != nil {
-		logger.Error(fmt.Sprintf("Unexpected error looking for preexisting bootstrap Secret: %s", err))
+		c.Log.Error(fmt.Sprintf("Unexpected error looking for preexisting bootstrap Secret: %s", err))
 		return 1
 	}
 
 	var updateServerPolicy bool
 	if bootstrapToken != "" {
-		logger.Info(fmt.Sprintf("ACLs already bootstrapped - retrieved bootstrap token from Secret %q", bootTokenSecretName))
+		c.Log.Info(fmt.Sprintf("ACLs already bootstrapped - retrieved bootstrap token from Secret %q", bootTokenSecretName))
 
 		// Mark that we should update the server ACL policy in case
 		// there are namespace related config changes. Because of the
@@ -248,18 +249,18 @@ func (c *Command) Run(args []string) int {
 		// otherwise won't be updated.
 		updateServerPolicy = true
 	} else {
-		logger.Info("No bootstrap token from previous installation found, continuing on to bootstrapping")
-		bootstrapToken, err = c.bootstrapServers(logger, bootTokenSecretName, scheme)
+		c.Log.Info("No bootstrap token from previous installation found, continuing on to bootstrapping")
+		bootstrapToken, err = c.bootstrapServers(bootTokenSecretName, scheme)
 		if err != nil {
-			logger.Error(err.Error())
+			c.Log.Error(err.Error())
 			return 1
 		}
 	}
 
 	// For all of the next operations we'll need a Consul client.
-	serverPods, err := c.getConsulServers(logger, 1, scheme)
+	serverPods, err := c.getConsulServers(1, scheme)
 	if err != nil {
-		logger.Error(err.Error())
+		c.Log.Error(err.Error())
 		return 1
 	}
 	serverAddr := serverPods[0].Addr
@@ -273,7 +274,7 @@ func (c *Command) Run(args []string) int {
 		},
 	})
 	if err != nil {
-		logger.Error(fmt.Sprintf("Error creating Consul client for addr %q: %s", serverAddr, err))
+		c.Log.Error(fmt.Sprintf("Error creating Consul client for addr %q: %s", serverAddr, err))
 		return 1
 	}
 
@@ -282,9 +283,9 @@ func (c *Command) Run(args []string) int {
 	// users upgrade to 1.7+. This updates the policy if the bootstrap
 	// token had previously existed, which signals a potential config change.
 	if updateServerPolicy {
-		_, err = c.setServerPolicy(consulClient, logger)
+		_, err = c.setServerPolicy(consulClient)
 		if err != nil {
-			logger.Error("Error updating the server ACL policy", "err", err)
+			c.Log.Error("Error updating the server ACL policy", "err", err)
 			return 1
 		}
 	}
@@ -292,21 +293,21 @@ func (c *Command) Run(args []string) int {
 	if c.flagCreateClientToken {
 		agentRules, err := c.agentRules()
 		if err != nil {
-			logger.Error("Error templating client agent rules", "err", err)
+			c.Log.Error("Error templating client agent rules", "err", err)
 			return 1
 		}
 
-		err = c.createACL("client", agentRules, consulClient, logger)
+		err = c.createACL("client", agentRules, consulClient)
 		if err != nil {
-			logger.Error(err.Error())
+			c.Log.Error(err.Error())
 			return 1
 		}
 	}
 
 	if c.flagAllowDNS {
-		err := c.configureDNSPolicies(logger, consulClient)
+		err := c.configureDNSPolicies(consulClient)
 		if err != nil {
-			logger.Error(err.Error())
+			c.Log.Error(err.Error())
 			return 1
 		}
 	}
@@ -314,13 +315,13 @@ func (c *Command) Run(args []string) int {
 	if c.flagCreateSyncToken {
 		syncRules, err := c.syncRules()
 		if err != nil {
-			logger.Error("Error templating sync rules", "err", err)
+			c.Log.Error("Error templating sync rules", "err", err)
 			return 1
 		}
 
-		err = c.createACL("catalog-sync", syncRules, consulClient, logger)
+		err = c.createACL("catalog-sync", syncRules, consulClient)
 		if err != nil {
-			logger.Error(err.Error())
+			c.Log.Error(err.Error())
 			return 1
 		}
 	}
@@ -328,29 +329,29 @@ func (c *Command) Run(args []string) int {
 	if c.flagCreateInjectToken {
 		injectRules, err := c.injectRules()
 		if err != nil {
-			logger.Error("Error templating inject rules", "err", err)
+			c.Log.Error("Error templating inject rules", "err", err)
 			return 1
 		}
 
-		err = c.createACL("connect-inject", injectRules, consulClient, logger)
+		err = c.createACL("connect-inject", injectRules, consulClient)
 		if err != nil {
-			logger.Error(err.Error())
+			c.Log.Error(err.Error())
 			return 1
 		}
 	}
 
 	if c.flagCreateEntLicenseToken {
-		err := c.createACL("enterprise-license", entLicenseRules, consulClient, logger)
+		err := c.createACL("enterprise-license", entLicenseRules, consulClient)
 		if err != nil {
-			logger.Error(err.Error())
+			c.Log.Error(err.Error())
 			return 1
 		}
 	}
 
 	if c.flagCreateSnapshotAgentToken {
-		err := c.createACL("client-snapshot-agent", snapshotAgentRules, consulClient, logger)
+		err := c.createACL("client-snapshot-agent", snapshotAgentRules, consulClient)
 		if err != nil {
-			logger.Error(err.Error())
+			c.Log.Error(err.Error())
 			return 1
 		}
 	}
@@ -358,33 +359,33 @@ func (c *Command) Run(args []string) int {
 	if c.flagCreateMeshGatewayToken {
 		meshGatewayRules, err := c.meshGatewayRules()
 		if err != nil {
-			logger.Error("Error templating dns rules", "err", err)
+			c.Log.Error("Error templating dns rules", "err", err)
 			return 1
 		}
 
-		err = c.createACL("mesh-gateway", meshGatewayRules, consulClient, logger)
+		err = c.createACL("mesh-gateway", meshGatewayRules, consulClient)
 		if err != nil {
-			logger.Error(err.Error())
+			c.Log.Error(err.Error())
 			return 1
 		}
 	}
 
 	if c.flagCreateInjectAuthMethod {
-		err := c.configureConnectInject(logger, consulClient)
+		err := c.configureConnectInject(consulClient)
 		if err != nil {
-			logger.Error(err.Error())
+			c.Log.Error(err.Error())
 			return 1
 		}
 	}
 
-	logger.Info("server-acl-init completed successfully")
+	c.Log.Info("server-acl-init completed successfully")
 	return 0
 }
 
 // getBootstrapToken returns the existing bootstrap token if there is one by
 // reading the Kubernetes Secret with name secretName.
 // If there is no bootstrap token yet, then it returns an empty string (not an error).
-func (c *Command) getBootstrapToken(logger hclog.Logger, secretName string) (string, error) {
+func (c *Command) getBootstrapToken(secretName string) (string, error) {
 	secret, err := c.clientset.CoreV1().Secrets(c.flagK8sNamespace).Get(secretName, metav1.GetOptions{})
 	if err != nil {
 		if k8serrors.IsNotFound(err) {
@@ -411,541 +412,17 @@ func (c *Command) configureKubeClient() error {
 	return nil
 }
 
-// getConsulServers returns n Consul server pods with their http addresses.
-// If there are less server pods than 'n' then the function will wait.
-func (c *Command) getConsulServers(logger hclog.Logger, n int, scheme string) ([]podAddr, error) {
-	var serverPods *apiv1.PodList
-	err := c.untilSucceeds("discovering Consul server pods",
-		func() error {
-			var err error
-			serverPods, err = c.clientset.CoreV1().Pods(c.flagK8sNamespace).List(metav1.ListOptions{LabelSelector: c.flagServerLabelSelector})
-			if err != nil {
-				return err
-			}
-
-			if len(serverPods.Items) == 0 {
-				return fmt.Errorf("no server pods with labels %q found", c.flagServerLabelSelector)
-			}
-
-			if len(serverPods.Items) < n {
-				return fmt.Errorf("found %d servers, require %d", len(serverPods.Items), n)
-			}
-
-			for _, pod := range serverPods.Items {
-				if pod.Status.PodIP == "" {
-					return fmt.Errorf("pod %s has no IP", pod.Name)
-				}
-			}
-			return nil
-		}, logger)
-	if err != nil {
-		return nil, err
-	}
-
-	var podAddrs []podAddr
-	for _, pod := range serverPods.Items {
-		var httpPort int32
-		for _, p := range pod.Spec.Containers[0].Ports {
-			if p.Name == scheme {
-				httpPort = p.ContainerPort
-			}
-		}
-		if httpPort == 0 {
-			return nil, fmt.Errorf("pod %s has no port labeled '%s'", pod.Name, scheme)
-		}
-		addr := fmt.Sprintf("%s:%d", pod.Status.PodIP, httpPort)
-		podAddrs = append(podAddrs, podAddr{
-			Name: pod.Name,
-			Addr: addr,
-		})
-	}
-	return podAddrs, nil
-}
-
-// bootstrapServers bootstraps ACLs and ensures each server has an ACL token.
-func (c *Command) bootstrapServers(logger hclog.Logger, bootTokenSecretName, scheme string) (string, error) {
-	serverPods, err := c.getConsulServers(logger, c.flagReplicas, scheme)
-	if err != nil {
-		return "", err
-	}
-	logger.Info(fmt.Sprintf("Found %d Consul server Pods", len(serverPods)))
-
-	// Pick the first pod to connect to for bootstrapping and set up connection.
-	firstServerAddr := serverPods[0].Addr
-	consulClient, err := api.NewClient(&api.Config{
-		Address: firstServerAddr,
-		Scheme:  scheme,
-		TLSConfig: api.TLSConfig{
-			Address: c.flagConsulTLSServerName,
-			CAFile:  c.flagConsulCACert,
-		},
-	})
-	if err != nil {
-		return "", fmt.Errorf("creating Consul client for address %s: %s", firstServerAddr, err)
-	}
-
-	// Call bootstrap ACLs API.
-	var bootstrapToken []byte
-	var unrecoverableErr error
-	err = c.untilSucceeds("bootstrapping ACLs - PUT /v1/acl/bootstrap",
-		func() error {
-			bootstrapResp, _, err := consulClient.ACL().Bootstrap()
-			if err == nil {
-				bootstrapToken = []byte(bootstrapResp.SecretID)
-				return nil
-			}
-
-			// Check if already bootstrapped.
-			if strings.Contains(err.Error(), "Unexpected response code: 403") {
-				unrecoverableErr = errors.New("ACLs already bootstrapped but the ACL token was not written to a Kubernetes secret." +
-					" We can't proceed because the bootstrap token is lost." +
-					" You must reset ACLs.")
-				return nil
-			}
-
-			if isNoLeaderErr(err) {
-				// Return a more descriptive error in the case of no leader
-				// being elected.
-				return fmt.Errorf("no leader elected: %s", err)
-			}
-			return err
-		}, logger)
-	if unrecoverableErr != nil {
-		return "", unrecoverableErr
-	}
-	if err != nil {
-		return "", err
-	}
-
-	// Write bootstrap token to a Kubernetes secret.
-	err = c.untilSucceeds(fmt.Sprintf("writing bootstrap Secret %q", bootTokenSecretName),
-		func() error {
-			secret := &apiv1.Secret{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: bootTokenSecretName,
-				},
-				Data: map[string][]byte{
-					"token": bootstrapToken,
-				},
-			}
-			_, err := c.clientset.CoreV1().Secrets(c.flagK8sNamespace).Create(secret)
-			return err
-		}, logger)
-	if err != nil {
-		return "", err
-	}
-
-	// Override our original client with a new one that has the bootstrap token
-	// set.
-	consulClient, err = api.NewClient(&api.Config{
-		Address: firstServerAddr,
-		Scheme:  scheme,
-		Token:   string(bootstrapToken),
-		TLSConfig: api.TLSConfig{
-			Address: c.flagConsulTLSServerName,
-			CAFile:  c.flagConsulCACert,
-		},
-	})
-	if err != nil {
-		return "", fmt.Errorf("creating Consul client for address %s: %s", firstServerAddr, err)
-	}
-
-	// Create new tokens for each server and apply them.
-	if err := c.setServerTokens(logger, consulClient, serverPods, string(bootstrapToken), scheme); err != nil {
-		return "", err
-	}
-	return string(bootstrapToken), nil
-}
-
-// setServerTokens creates policies and associated ACL token for each server
-// and then provides the token to the server.
-func (c *Command) setServerTokens(logger hclog.Logger, consulClient *api.Client,
-	serverPods []podAddr, bootstrapToken, scheme string) error {
-
-	agentPolicy, err := c.setServerPolicy(consulClient, logger)
-	if err != nil {
-		return err
-	}
-
-	// Create agent token for each server agent.
-	var serverTokens []api.ACLToken
-	for _, pod := range serverPods {
-		var token *api.ACLToken
-		err := c.untilSucceeds(fmt.Sprintf("creating server token for %s - PUT /v1/acl/token", pod.Name),
-			func() error {
-				tokenReq := api.ACLToken{
-					Description: fmt.Sprintf("Server Token for %s", pod.Name),
-					Policies:    []*api.ACLTokenPolicyLink{{Name: agentPolicy.Name}},
-				}
-				var err error
-				token, _, err = consulClient.ACL().TokenCreate(&tokenReq, nil)
-				return err
-			}, logger)
-		if err != nil {
-			return err
-		}
-		serverTokens = append(serverTokens, *token)
-	}
-
-	// Pass out agent tokens to servers.
-	for i, pod := range serverPods {
-		// We create a new client for each server because we need to call each
-		// server specifically.
-		serverClient, err := api.NewClient(&api.Config{
-			Address: pod.Addr,
-			Scheme:  scheme,
-			Token:   bootstrapToken,
-			TLSConfig: api.TLSConfig{
-				Address: c.flagConsulTLSServerName,
-				CAFile:  c.flagConsulCACert,
-			},
-		})
-		if err != nil {
-			return fmt.Errorf(" creating Consul client for address %q: %s", pod.Addr, err)
-		}
-		podName := pod.Name
-
-		// Update token.
-		err = c.untilSucceeds(fmt.Sprintf("updating server token for %s - PUT /v1/agent/token/agent", podName),
-			func() error {
-				_, err := serverClient.Agent().UpdateAgentACLToken(serverTokens[i].SecretID, nil)
-				return err
-			}, logger)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (c *Command) setServerPolicy(consulClient *api.Client, logger hclog.Logger) (api.ACLPolicy, error) {
-	agentRules, err := c.agentRules()
-	if err != nil {
-		logger.Error("Error templating server agent rules", "err", err)
-		return api.ACLPolicy{}, err
-	}
-
-	// Create agent policy.
-	agentPolicy := api.ACLPolicy{
-		Name:        "agent-token",
-		Description: "Agent Token Policy",
-		Rules:       agentRules,
-	}
-	err = c.untilSucceeds("creating agent policy - PUT /v1/acl/policy",
-		func() error {
-			return c.createOrUpdateACLPolicy(agentPolicy, logger, consulClient)
-		}, logger)
-	if err != nil {
-		return api.ACLPolicy{}, err
-	}
-
-	return agentPolicy, nil
-}
-
-// createACL creates a policy with rules and name, creates an ACL token for that
-// policy and then writes the token to a Kubernetes secret.
-func (c *Command) createACL(name, rules string, consulClient *api.Client, logger hclog.Logger) error {
-	// Create policy with the given rules.
-	policyTmpl := api.ACLPolicy{
-		Name:        fmt.Sprintf("%s-token", name),
-		Description: fmt.Sprintf("%s Token Policy", name),
-		Rules:       rules,
-	}
-	err := c.untilSucceeds(fmt.Sprintf("creating %s policy", policyTmpl.Name),
-		func() error {
-			return c.createOrUpdateACLPolicy(policyTmpl, logger, consulClient)
-		}, logger)
-	if err != nil {
-		return err
-	}
-
-	// Check if the secret already exists, if so, we assume the ACL has already been
-	// created and return.
-	secretName := c.withPrefix(name + "-acl-token")
-	_, err = c.clientset.CoreV1().Secrets(c.flagK8sNamespace).Get(secretName, metav1.GetOptions{})
-	if err == nil {
-		logger.Info(fmt.Sprintf("Secret %q already exists", secretName))
-		return nil
-	}
-
-	// Create token for the policy if the secret did not exist previously.
-	tokenTmpl := api.ACLToken{
-		Description: fmt.Sprintf("%s Token", name),
-		Policies:    []*api.ACLTokenPolicyLink{{Name: policyTmpl.Name}},
-	}
-	var token string
-	err = c.untilSucceeds(fmt.Sprintf("creating token for policy %s", policyTmpl.Name),
-		func() error {
-			createdToken, _, err := consulClient.ACL().TokenCreate(&tokenTmpl, &api.WriteOptions{})
-			if err == nil {
-				token = createdToken.SecretID
-			}
-			return err
-		}, logger)
-	if err != nil {
-		return err
-	}
-
-	// Write token to a Kubernetes secret.
-	return c.untilSucceeds(fmt.Sprintf("writing Secret for token %s", policyTmpl.Name),
-		func() error {
-			secret := &apiv1.Secret{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: secretName,
-				},
-				Data: map[string][]byte{
-					"token": []byte(token),
-				},
-			}
-			_, err := c.clientset.CoreV1().Secrets(c.flagK8sNamespace).Create(secret)
-			return err
-		}, logger)
-}
-
-// configureDNSPolicies sets up policies and tokens so that Consul DNS will
-// work.
-func (c *Command) configureDNSPolicies(logger hclog.Logger, consulClient *api.Client) error {
-	dnsRules, err := c.dnsRules()
-	if err != nil {
-		logger.Error("Error templating dns rules", "err", err)
-		return err
-	}
-
-	// Create policy for the anonymous token
-	dnsPolicy := api.ACLPolicy{
-		Name:        "dns-policy",
-		Description: "DNS Policy",
-		Rules:       dnsRules,
-	}
-
-	err = c.untilSucceeds("creating dns policy - PUT /v1/acl/policy",
-		func() error {
-			return c.createOrUpdateACLPolicy(dnsPolicy, logger, consulClient)
-		}, logger)
-	if err != nil {
-		return err
-	}
-
-	// Create token to get sent to TokenUpdate
-	aToken := api.ACLToken{
-		AccessorID: "00000000-0000-0000-0000-000000000002",
-		Policies:   []*api.ACLTokenPolicyLink{{Name: dnsPolicy.Name}},
-	}
-
-	// Update anonymous token to include this policy
-	return c.untilSucceeds("updating anonymous token with DNS policy",
-		func() error {
-			_, _, err := consulClient.ACL().TokenUpdate(&aToken, &api.WriteOptions{})
-			return err
-		}, logger)
-}
-
-// configureConnectInject sets up auth methods so that connect injection will
-// work.
-func (c *Command) configureConnectInject(logger hclog.Logger, consulClient *api.Client) error {
-	// First, check if there's already an acl binding rule. If so, then this
-	// work is already done.
-	authMethodName := c.withPrefix("k8s-auth-method")
-	var existingRules []*api.ACLBindingRule
-	err := c.untilSucceeds(fmt.Sprintf("listing binding rules for auth method %s", authMethodName),
-		func() error {
-			var err error
-			existingRules, _, err = consulClient.ACL().BindingRuleList(authMethodName, nil)
-			return err
-		}, logger)
-	if err != nil {
-		return err
-	}
-	// If Consul namespaces are enabled, someone may have changed their namespace
-	// configuration, so auth methods and binding rules should be updated accordingly
-	if len(existingRules) > 0 && !c.flagEnableNamespaces {
-		logger.Info(fmt.Sprintf("Binding rule for %s already exists", authMethodName))
-		return nil
-	}
-
-	var kubeSvc *apiv1.Service
-	err = c.untilSucceeds("getting kubernetes service IP",
-		func() error {
-			var err error
-			kubeSvc, err = c.clientset.CoreV1().Services("default").Get("kubernetes", metav1.GetOptions{})
-			return err
-		}, logger)
-	if err != nil {
-		return err
-	}
-
-	// Get the Secret name for the auth method ServiceAccount.
-	var authMethodServiceAccount *apiv1.ServiceAccount
-	saName := c.withPrefix("connect-injector-authmethod-svc-account")
-	err = c.untilSucceeds(fmt.Sprintf("getting %s ServiceAccount", saName),
-		func() error {
-			var err error
-			authMethodServiceAccount, err = c.clientset.CoreV1().ServiceAccounts(c.flagK8sNamespace).Get(saName, metav1.GetOptions{})
-			return err
-		}, logger)
-	if err != nil {
-		return err
-	}
-
-	// ServiceAccounts always have a secret name. The secret
-	// contains the JWT token.
-	saSecretName := authMethodServiceAccount.Secrets[0].Name
-
-	// Get the secret that will contain the ServiceAccount JWT token.
-	var saSecret *apiv1.Secret
-	err = c.untilSucceeds(fmt.Sprintf("getting %s Secret", saSecretName),
-		func() error {
-			var err error
-			saSecret, err = c.clientset.CoreV1().Secrets(c.flagK8sNamespace).Get(saSecretName, metav1.GetOptions{})
-			return err
-		}, logger)
-	if err != nil {
-		return err
-	}
-
-	// Now we're ready to set up Consul's auth method.
-	authMethodTmpl := api.ACLAuthMethod{
-		Name:        authMethodName,
-		Description: "Kubernetes AuthMethod",
-		Type:        "kubernetes",
-		Config: map[string]interface{}{
-			"Host":              fmt.Sprintf("https://%s:443", kubeSvc.Spec.ClusterIP),
-			"CACert":            string(saSecret.Data["ca.crt"]),
-			"ServiceAccountJWT": string(saSecret.Data["token"]),
-		},
-	}
-
-	// Add options for mirroring namespaces
-	if c.flagEnableInjectK8SNSMirroring {
-		authMethodTmpl.Config["MapNamespaces"] = true
-		authMethodTmpl.Config["ConsulNamespacePrefix"] = c.flagInjectK8SNSMirroringPrefix
-	}
-
-	// Set up the auth method in the specific namespace if not mirroring
-	// If namespaces and mirroring are enabled, this is not necessary because
-	// the auth method will fall back to being created in the Consul `default`
-	// namespace automatically, as is necessary for mirroring.
-	// Note: if the config changes, an auth method will be created in the
-	// correct namespace, but the old auth method will not be removed.
-	writeOptions := api.WriteOptions{}
-	if c.flagEnableNamespaces && !c.flagEnableInjectK8SNSMirroring {
-		writeOptions.Namespace = c.flagConsulInjectDestinationNamespace
-	}
-
-	var authMethod *api.ACLAuthMethod
-	err = c.untilSucceeds(fmt.Sprintf("creating auth method %s", authMethodTmpl.Name),
-		func() error {
-			var err error
-			// `AuthMethodCreate` will also be able to update an existing
-			// AuthMethod based on the name provided. This means that any namespace
-			// configuration changes will correctly update the AuthMethod.
-			authMethod, _, err = consulClient.ACL().AuthMethodCreate(&authMethodTmpl, &writeOptions)
-			return err
-		}, logger)
-	if err != nil {
-		return err
-	}
-
-	// Create the binding rule.
-	abr := api.ACLBindingRule{
-		Description: "Kubernetes binding rule",
-		AuthMethod:  authMethod.Name,
-		BindType:    api.BindingRuleBindTypeService,
-		BindName:    "${serviceaccount.name}",
-		Selector:    c.flagBindingRuleSelector,
-	}
-
-	// Add a namespace if appropriate
-	// If namespaces and mirroring are enabled, this is not necessary because
-	// the binding rule will fall back to being created in the Consul `default`
-	// namespace automatically, as is necessary for mirroring.
-	if c.flagEnableNamespaces && !c.flagEnableInjectK8SNSMirroring {
-		abr.Namespace = c.flagConsulInjectDestinationNamespace
-	}
-
-	// If the binding rule already exists and namespaces are enabled, update it
-	if len(existingRules) > 0 && c.flagEnableNamespaces {
-		// Find the policy that matches our name and description
-		// and that's the ID we need
-		for _, existingRule := range existingRules {
-			if existingRule.BindName == abr.BindName && existingRule.Description == abr.Description {
-				abr.ID = existingRule.ID
-			}
-		}
-
-		// This will only happen if there are existing policies
-		// for this auth method, but none that match the binding
-		// rule set up here in the bootstrap method.
-		if abr.ID == "" {
-			return errors.New("Unable to find a matching ACL binding rule to update")
-		}
-
-		err = c.untilSucceeds(fmt.Sprintf("updating acl binding rule for %s", authMethodTmpl.Name),
-			func() error {
-				_, _, err := consulClient.ACL().BindingRuleUpdate(&abr, nil)
-				return err
-			}, logger)
-	} else {
-		// Otherwise create the binding rule
-		err = c.untilSucceeds(fmt.Sprintf("creating acl binding rule for %s", authMethodTmpl.Name),
-			func() error {
-				_, _, err := consulClient.ACL().BindingRuleCreate(&abr, nil)
-				return err
-			}, logger)
-	}
-	return err
-}
-
-func (c *Command) createOrUpdateACLPolicy(policy api.ACLPolicy, logger hclog.Logger, consulClient *api.Client) error {
-	// Attempt to create the ACL policy
-	_, _, err := consulClient.ACL().PolicyCreate(&policy, &api.WriteOptions{})
-
-	// With the introduction of Consul namespaces, if someone upgrades into a
-	// Consul version with namespace support or changes any of their namespace
-	// settings, the policies associated with their ACL tokens will need to be
-	// updated to be namespace aware.
-	if isPolicyExistsErr(err, policy.Name) {
-		logger.Info(fmt.Sprintf("Policy %q already exists, updating", policy.Name))
-
-		// The policy ID is required in any PolicyUpdate call, so first we need to
-		// get the existing policy to extract its ID.
-		existingPolicies, _, err := consulClient.ACL().PolicyList(&api.QueryOptions{})
-		if err != nil {
-			return err
-		}
-
-		// Find the policy that matches our name and description
-		// and that's the ID we need
-		for _, existingPolicy := range existingPolicies {
-			if existingPolicy.Name == policy.Name && existingPolicy.Description == policy.Description {
-				policy.ID = existingPolicy.ID
-			}
-		}
-
-		// This shouldn't happen, because we're looking for a policy
-		// only after we've hit a `Policy already exists` error.
-		if policy.ID == "" {
-			return errors.New("Unable to find existing ACL policy")
-		}
-
-		// Update the policy now that we've found its ID
-		_, _, err = consulClient.ACL().PolicyUpdate(&policy, &api.WriteOptions{})
-		return err
-	}
-	return err
-}
-
 // untilSucceeds runs op until it returns a nil error.
 // If c.cmdTimeout is cancelled it will exit.
-func (c *Command) untilSucceeds(opName string, op func() error, logger hclog.Logger) error {
+func (c *Command) untilSucceeds(opName string, op func() error) error {
 	for {
 		err := op()
 		if err == nil {
-			logger.Info(fmt.Sprintf("Success: %s", opName))
+			c.Log.Info(fmt.Sprintf("Success: %s", opName))
 			break
 		}
-		logger.Error(fmt.Sprintf("Failure: %s", opName), "err", err)
-		logger.Info("Retrying in " + c.retryDuration.String())
+		c.Log.Error(fmt.Sprintf("Failure: %s", opName), "err", err)
+		c.Log.Info("Retrying in " + c.retryDuration.String())
 		// Wait on either the retry duration (in which case we continue) or the
 		// overall command timeout.
 		select {
@@ -968,30 +445,6 @@ func (c *Command) withPrefix(resource string) string {
 	// the -release-name flag. We ensure that this is set if -resource-prefix
 	// is not set when parsing the flags.
 	return fmt.Sprintf("%s-consul-%s", c.flagReleaseName, resource)
-}
-
-// isNoLeaderErr returns true if err is due to trying to call the
-// bootstrap ACLs API when there is no leader elected.
-func isNoLeaderErr(err error) bool {
-	return err != nil && strings.Contains(err.Error(), "Unexpected response code: 500") &&
-		strings.Contains(err.Error(), "The ACL system is currently in legacy mode.")
-}
-
-// isPolicyExistsErr returns true if err is due to trying to call the
-// policy create API when the policy already exists.
-func isPolicyExistsErr(err error, policyName string) bool {
-	return err != nil &&
-		strings.Contains(err.Error(), "Unexpected response code: 500") &&
-		strings.Contains(err.Error(), fmt.Sprintf("Invalid Policy: A Policy with Name %q already exists", policyName))
-}
-
-// podAddr is a convenience struct for passing around pod names and
-// addresses for Consul servers.
-type podAddr struct {
-	// Name is the name of the pod.
-	Name string
-	// Addr is in the form "<ip>:<port>".
-	Addr string
 }
 
 const synopsis = "Initialize ACLs on Consul servers and other components."
